refactor lesson editor into hook and components
